refactor: create test files structure
