+++ src/components/MoviesList/MoviesList.tsx
@@ -1,6 +1,6 @@
-import type { MoviePoster } from "../api/api";
+import type { MoviePoster } from "../../api/api";
 
-import MovieCard from "./MovieCard";
+import MovieCard from "../MovieCard/MovieCard";
 
 export default function MoviesList({ movies }: { movies: MoviePoster[] }) {
   return (
